Follow the spec: the mpris version is two uint16 fields, not two int fields.
svn path=/trunk/extragear/multimedia/amarok/; revision=845785

+++ src/dbus/RootDBusHandler.h
@@ -24,8 +24,8 @@
 
 struct Version
 {
-    int major;
-    int minor;
+    quint16 major;
+    quint16 minor;
 };
 
 Q_DECLARE_METATYPE(Version)
